# contracts/openapi/teamos.yaml
openapi: 3.1.0
...
info:
  title: TeamOS Gateway API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    Внешний JSON/REST-контракт TeamOS. Поля и формы ответов зеркалят
    `src/api/index.ts` и `src/types/index.ts` frontend-приложения.
    Все даты-время передаются в ISO-8601, идентификаторы являются UUID.
servers:
  - url: http://localhost:8080
    description: Локальный gateway
...
security:
  - bearerAuth: []
paths:
  /api/v1/auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Получить текущего пользователя
      responses:
        '200':
          $ref: '#/components/responses/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Auth]
      operationId: updateCurrentUser
      summary: Обновить профиль текущего пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCurrentUserInput'
      responses:
        '200':
          $ref: '#/components/responses/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
...
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refresh
      summary: Ротировать refresh-сессию и выдать новый access-токен
      security:
        - refreshCookie: []
      responses:
        '200':
          $ref: '#/components/responses/AuthSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
...
  /api/v1/auth/invites/{token}:
    parameters:
      - $ref: '#/components/parameters/InviteToken'
    get:
      tags: [Auth]
      operationId: getInviteByToken
      summary: Получить ожидающее приглашение по токену
      security: []
      responses:
        '200':
          $ref: '#/components/responses/Invite'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/auth/invites/{token}/accept:
    parameters:
      - $ref: '#/components/parameters/InviteToken'
    post:
      tags: [Auth]
      operationId: acceptInvite
      summary: Принять приглашение и задать пароль
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptInviteInput'
      responses:
        '200':
          $ref: '#/components/responses/AuthSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/company:
    get:
      tags: [Company]
      operationId: getCompany
      summary: Получить компанию текущего пользователя
      responses:
        '200':
          $ref: '#/components/responses/Company'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Company]
      operationId: updateCompany
      summary: Обновить компанию
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCompanyInput'
      responses:
        '200':
          $ref: '#/components/responses/Company'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/org/departments:
    get:
      tags: [Organization]
      operationId: getDepartments
      summary: Получить отделы
      responses:
        '200':
          $ref: '#/components/responses/Departments'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Organization]
      operationId: createDepartment
      summary: Создать отдел
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepartmentInput'
      responses:
        '201':
          $ref: '#/components/responses/Department'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/org/departments/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    patch:
      tags: [Organization]
      operationId: updateDepartment
      summary: Обновить отдел
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDepartmentInput'
      responses:
        '200':
          $ref: '#/components/responses/Department'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Organization]
      operationId: deleteDepartment
      summary: Удалить пустой отдел
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/org/departments/{id}/move:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Organization]
      operationId: moveDepartment
      summary: Переместить отдел
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveDepartmentInput'
      responses:
        '200':
          $ref: '#/components/responses/Department'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/org/positions:
    get:
      tags: [Organization]
      operationId: getPositions
      summary: Получить должности
      responses:
        '200':
          $ref: '#/components/responses/Positions'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Organization]
      operationId: createPosition
      summary: Создать должность
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePositionInput'
      responses:
        '201':
          $ref: '#/components/responses/Position'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/org/positions/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Organization]
      operationId: getPosition
      summary: Получить должность
      responses:
        '200':
          $ref: '#/components/responses/Position'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Organization]
      operationId: updatePosition
      summary: Обновить должность, включая перенос в другой отдел
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePositionInput'
      responses:
        '200':
          $ref: '#/components/responses/Position'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Organization]
      operationId: deletePosition
      summary: Удалить должность и снять её с сотрудников
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/org/positions/{id}/move:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Organization]
      operationId: movePosition
      summary: Переместить должность в другой отдел
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovePositionInput'
      responses:
        '200':
          $ref: '#/components/responses/Position'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/org/users:
    get:
      tags: [Organization]
      operationId: getUsers
      summary: Получить сотрудников
      responses:
        '200':
          $ref: '#/components/responses/Users'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Organization]
      operationId: createUser
      summary: Создать активного сотрудника
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserInput'
      responses:
        '201':
          $ref: '#/components/responses/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/org/users/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Organization]
      operationId: getUser
      summary: Получить сотрудника
      responses:
        '200':
          $ref: '#/components/responses/User'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Organization]
      operationId: updateUser
      summary: Обновить сотрудника
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserInput'
      responses:
        '200':
          $ref: '#/components/responses/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/org/invites:
    get:
      tags: [Organization]
      operationId: getInvites
      summary: Получить приглашения, новые первыми
      responses:
        '200':
          $ref: '#/components/responses/Invites'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Organization]
      operationId: inviteUser
      summary: Создать приглашение по email или ссылке
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteUserInput'
      responses:
        '201':
          $ref: '#/components/responses/Invite'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: teamos_refresh
...
  parameters:
...
    InviteToken:
      name: token
      in: path
      required: true
      schema:
        type: string
        minLength: 1
  responses:
...
    BadRequest:
      description: Ошибка валидации или доменный запрет
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Учётные данные отсутствуют или недействительны
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Error:
      description: Ошибка API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    User:
      description: Пользователь
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/User'
...
  schemas:
...
    UpdateCurrentUserInput:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        phone:
          type: string
        avatarUrl:
          type: string
...